Improves handling ```size_t``` FFI values.
- The size_t type that is used in all the MU APIs will
  be dependent on the compiler that ```bindgen``` uses to
  generate the bindings for the library. This attempts to
  make the crate less dependant on that value in the
  public APIs.

- Changes the Marshall and Unmarshall traits to use usize as
  input.

- Removed the ```private_in_public``` it is 'warn by default' and
  because documentation ci is run with ```-Dwarnings``` it will
  cause the lint to be enforced.

+++ tss-esapi/tests/integration_tests/common/marshall.rs
@@ -19,20 +19,20 @@ pub fn check_marshall_unmarshall_offset<T: Marshall + UnMarshall + Eq + std::fmt
 
     val.marshall_offset(&mut buf, &mut offset)
         .expect("Failed first marshall_offset");
-    assert_eq!(offset, len as u64);
+    assert_eq!(offset, len);
 
     val.marshall_offset(&mut buf, &mut offset)
         .expect("Failed second marshall_offset");
-    assert_eq!(offset, (len * 2) as u64);
+    assert_eq!(offset, (len * 2));
 
     offset = 0;
     let unmarshalled_one =
         T::unmarshall_offset(&buf, &mut offset).expect("Failed to unmarshall_offset first copy");
-    assert_eq!(offset, len as u64);
+    assert_eq!(offset, len);
     assert_eq!(val, &unmarshalled_one);
 
     let unmarshalled_two =
         T::unmarshall_offset(&buf, &mut offset).expect("Failed to unmarshall_offset second copy");
-    assert_eq!(offset, (len * 2) as u64);
+    assert_eq!(offset, (len * 2));
     assert_eq!(val, &unmarshalled_two);
 }
